Changed some return types

+++ Student.h
@@ -18,9 +18,9 @@ public:
     string GetFirstName() const;
     string GetLastName() const;
     string GetCourse() const;
-    virtual int GetGradeAvg() const=0;
+    virtual float GetGradeAvg() const=0;
     virtual char GetLetterGrade() const=0;
-    virtual int GetFinalExam() const=0;
+    virtual float GetFinalExam() const=0;
